requirements.txt
python-dotenv~=0.19.0
python-telegram-bot~=13.7
azure-cognitiveservices-speech~=1.19.0
pydub~=0.25.1